Split out getThemeCustomizeUrl
